Changed namespace to republish on Play, because initial account was removed
