+++ app/src/main/java/com/github/richoux/les_sons_en_francais/ui/OptionsData.kt
@@ -1,0 +1,8 @@
+package com.github.richoux.les_sons_en_francais.ui
+
+data class OptionsData(
+    val uppercase: Boolean = true,
+    val lowercase: Boolean = true,
+    val cursive: Boolean = true,
+    val darkMode: Boolean = false
+)
