+++ app/src/main/java/com/github/richoux/les_sons_en_francais/ui/theme/Color.kt
@@ -1,0 +1,9 @@
+package com.github.richoux.les_sons_en_francais.ui.theme
+
+import androidx.compose.ui.graphics.Color
+
+val Black = Color(0xFF000000)
+val White = Color(0xFFFFFFFF)
+val BackgroundTwitter = Color(0xFF15202B)
+val BlueTwitter = Color(0xFF1D9BF0)
+val LightBlue = Color(0xFF79C4F7)
