--- app/src/main/java/fr/richoux/lessonsenfrancais/PlayerFactory.kt
@@ -1,129 +1,0 @@
-package fr.richoux.lessonsenfrancais
-
-import android.content.Context
-import android.media.MediaPlayer
-
-// This function is plain stupid, but for some reason, the app crashed in release mode
-// when creating a MediaPlayer object on some sounds, like 'u' for instance.
-// This is the only way I found to get around the problem.
-fun PlayerFactory(
-    context: Context,
-    id: Int
-): MediaPlayer {
-    when( id ) {
-        R.raw.ail -> return MediaPlayer.create( context, R.raw.ail )
-        R.raw.ai -> return MediaPlayer.create( context, R.raw.ai )
-        R.raw.a -> return MediaPlayer.create( context, R.raw.a )
-        R.raw.an -> return MediaPlayer.create( context, R.raw.an )
-        R.raw.ba -> return MediaPlayer.create( context, R.raw.ba )
-        R.raw.be -> return MediaPlayer.create( context, R.raw.be )
-        R.raw.bi -> return MediaPlayer.create( context, R.raw.bi )
-        R.raw.bo -> return MediaPlayer.create( context, R.raw.bo )
-        R.raw.bu -> return MediaPlayer.create( context, R.raw.bu )
-        R.raw.cha -> return MediaPlayer.create( context, R.raw.cha )
-        R.raw.che -> return MediaPlayer.create( context, R.raw.che )
-        R.raw.chi -> return MediaPlayer.create( context, R.raw.chi )
-        R.raw.cho -> return MediaPlayer.create( context, R.raw.cho )
-        R.raw.chu -> return MediaPlayer.create( context, R.raw.chu )
-        R.raw.da -> return MediaPlayer.create( context, R.raw.da )
-        R.raw.de -> return MediaPlayer.create( context, R.raw.de )
-        R.raw.di -> return MediaPlayer.create( context, R.raw.di )
-        R.raw.doo -> return MediaPlayer.create( context, R.raw.doo )
-        R.raw.du -> return MediaPlayer.create( context, R.raw.du )
-        R.raw.eil -> return MediaPlayer.create( context, R.raw.eil )
-        R.raw.elle -> return MediaPlayer.create( context, R.raw.elle )
-        R.raw.e -> return MediaPlayer.create( context, R.raw.e )
-        R.raw.erre -> return MediaPlayer.create( context, R.raw.erre )
-        R.raw.esse -> return MediaPlayer.create( context, R.raw.esse )
-        R.raw.et -> return MediaPlayer.create( context, R.raw.et )
-        R.raw.ette -> return MediaPlayer.create( context, R.raw.ette )
-        R.raw.euil -> return MediaPlayer.create( context, R.raw.euil )
-        R.raw.fa -> return MediaPlayer.create( context, R.raw.fa )
-        R.raw.fe -> return MediaPlayer.create( context, R.raw.fe )
-        R.raw.fi -> return MediaPlayer.create( context, R.raw.fi )
-        R.raw.fo -> return MediaPlayer.create( context, R.raw.fo )
-        R.raw.fu -> return MediaPlayer.create( context, R.raw.fu )
-        R.raw.ga -> return MediaPlayer.create( context, R.raw.ga )
-        R.raw.gn -> return MediaPlayer.create( context, R.raw.gn )
-        R.raw.go -> return MediaPlayer.create( context, R.raw.go )
-        R.raw.gue -> return MediaPlayer.create( context, R.raw.gue )
-        R.raw.gui -> return MediaPlayer.create( context, R.raw.gui )
-        R.raw.gu -> return MediaPlayer.create( context, R.raw.gu )
-        R.raw.ien -> return MediaPlayer.create( context, R.raw.ien )
-        R.raw.ier -> return MediaPlayer.create( context, R.raw.ier )
-        R.raw.ill -> return MediaPlayer.create( context, R.raw.ill )
-        R.raw.i -> return MediaPlayer.create( context, R.raw.i )
-        R.raw.inn -> return MediaPlayer.create( context, R.raw.inn )
-        R.raw.ja -> return MediaPlayer.create( context, R.raw.ja )
-        R.raw.je -> return MediaPlayer.create( context, R.raw.je )
-        R.raw.ji -> return MediaPlayer.create( context, R.raw.ji )
-        R.raw.jo -> return MediaPlayer.create( context, R.raw.jo )
-        R.raw.ju -> return MediaPlayer.create( context, R.raw.ju )
-        R.raw.ka -> return MediaPlayer.create( context, R.raw.ka )
-        R.raw.ke -> return MediaPlayer.create( context, R.raw.ke )
-        R.raw.ki -> return MediaPlayer.create( context, R.raw.ki )
-        R.raw.ko -> return MediaPlayer.create( context, R.raw.ko )
-        R.raw.ku -> return MediaPlayer.create( context, R.raw.ku )
-        R.raw.la -> return MediaPlayer.create( context, R.raw.la )
-        R.raw.le -> return MediaPlayer.create( context, R.raw.le )
-        R.raw.li -> return MediaPlayer.create( context, R.raw.li )
-        R.raw.lo -> return MediaPlayer.create( context, R.raw.lo )
-        R.raw.lu -> return MediaPlayer.create( context, R.raw.lu )
-        R.raw.ma -> return MediaPlayer.create( context, R.raw.ma )
-        R.raw.me -> return MediaPlayer.create( context, R.raw.me )
-        R.raw.mi -> return MediaPlayer.create( context, R.raw.mi )
-        R.raw.mo -> return MediaPlayer.create( context, R.raw.mo )
-        R.raw.mu -> return MediaPlayer.create( context, R.raw.mu )
-        R.raw.na -> return MediaPlayer.create( context, R.raw.na )
-        R.raw.ne -> return MediaPlayer.create( context, R.raw.ne )
-        R.raw.ni -> return MediaPlayer.create( context, R.raw.ni )
-        R.raw.no -> return MediaPlayer.create( context, R.raw.no )
-        R.raw.nu -> return MediaPlayer.create( context, R.raw.nu )
-        R.raw.oi -> return MediaPlayer.create( context, R.raw.oi )
-        R.raw.oin -> return MediaPlayer.create( context, R.raw.oin )
-        R.raw.o -> return MediaPlayer.create( context, R.raw.o )
-        R.raw.on -> return MediaPlayer.create( context, R.raw.on )
-        R.raw.ouil -> return MediaPlayer.create( context, R.raw.ouil )
-        R.raw.ou -> return MediaPlayer.create( context, R.raw.ou )
-        R.raw.pa -> return MediaPlayer.create( context, R.raw.pa )
-        R.raw.pe -> return MediaPlayer.create( context, R.raw.pe )
-        R.raw.ph -> return MediaPlayer.create( context, R.raw.ph )
-        R.raw.pi -> return MediaPlayer.create( context, R.raw.pi )
-        R.raw.po -> return MediaPlayer.create( context, R.raw.po )
-        R.raw.pu -> return MediaPlayer.create( context, R.raw.pu )
-        R.raw.ra -> return MediaPlayer.create( context, R.raw.ra )
-        R.raw.re -> return MediaPlayer.create( context, R.raw.re )
-        R.raw.ri -> return MediaPlayer.create( context, R.raw.ri )
-        R.raw.ro -> return MediaPlayer.create( context, R.raw.ro )
-        R.raw.ru -> return MediaPlayer.create( context, R.raw.ru )
-        R.raw.sa -> return MediaPlayer.create( context, R.raw.sa )
-        R.raw.se -> return MediaPlayer.create( context, R.raw.se )
-        R.raw.si -> return MediaPlayer.create( context, R.raw.si )
-        R.raw.so -> return MediaPlayer.create( context, R.raw.so )
-        R.raw.su -> return MediaPlayer.create( context, R.raw.su )
-        R.raw.ta -> return MediaPlayer.create( context, R.raw.ta )
-        R.raw.te -> return MediaPlayer.create( context, R.raw.te )
-        R.raw.ti -> return MediaPlayer.create( context, R.raw.ti )
-        R.raw.tion -> return MediaPlayer.create( context, R.raw.tion )
-        R.raw.to -> return MediaPlayer.create( context, R.raw.to )
-        R.raw.tu -> return MediaPlayer.create( context, R.raw.tu )
-        R.raw.u -> return MediaPlayer.create( context, R.raw.u )
-        R.raw.une -> return MediaPlayer.create( context, R.raw.une )
-        R.raw.va -> return MediaPlayer.create( context, R.raw.va )
-        R.raw.ve -> return MediaPlayer.create( context, R.raw.ve )
-        R.raw.vi -> return MediaPlayer.create( context, R.raw.vi )
-        R.raw.vo -> return MediaPlayer.create( context, R.raw.vo )
-        R.raw.vu -> return MediaPlayer.create( context, R.raw.vu )
-        R.raw.ya -> return MediaPlayer.create( context, R.raw.ya )
-        R.raw.ye -> return MediaPlayer.create( context, R.raw.ye )
-        R.raw.yi -> return MediaPlayer.create( context, R.raw.yi )
-        R.raw.yo -> return MediaPlayer.create( context, R.raw.yo )
-        R.raw.yu -> return MediaPlayer.create( context, R.raw.yu )
-        R.raw.za -> return MediaPlayer.create( context, R.raw.za )
-        R.raw.ze -> return MediaPlayer.create( context, R.raw.ze )
-        R.raw.zi -> return MediaPlayer.create( context, R.raw.zi )
-        R.raw.zo -> return MediaPlayer.create( context, R.raw.zo )
-        R.raw.zu -> return MediaPlayer.create( context, R.raw.zu )
-        else -> { return MediaPlayer.create( context, R.raw.o ) }
-    }
-}
