--- app/src/main/java/fr/richoux/lessonsenfrancais/ui/OptionsData.kt
@@ -1,8 +1,0 @@
-package fr.richoux.lessonsenfrancais.ui
-
-data class OptionsData(
-    val uppercase: Boolean = true,
-    val lowercase: Boolean = true,
-    val cursive: Boolean = true,
-    val darkMode: Boolean = false
-)
