--- app/src/main/java/fr/richoux/lessonsenfrancais/ui/theme/Type.kt
@@ -1,27 +1,0 @@
-package fr.richoux.lessonsenfrancais.ui.theme
-
-import androidx.compose.material.Typography
-import androidx.compose.ui.text.TextStyle
-import androidx.compose.ui.text.font.Font
-import androidx.compose.ui.text.font.FontFamily
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.unit.sp
-import fr.richoux.lessonsenfrancais.R
-
-val Mulish = FontFamily(
-    Font(R.font.mulish_regular)
-)
-
-// Set of Material typography styles to start with
-val Typography = Typography(
-    body1 = TextStyle(
-        fontFamily = Mulish,
-        fontWeight = FontWeight.Normal,
-        fontSize = 16.sp
-    ),
-    body2 = TextStyle(
-        fontFamily = FontFamily.Cursive,
-        fontWeight = FontWeight.Normal,
-        fontSize = 12.sp
-    )
-)
