+++ fastlane/Appfile
@@ -1,0 +1,3 @@
+json_key_file("/Users/nicolasraoul/src/fastlane_playstore_key.json")
+package_name("com.github.richoux.les_sons_en_francais")
+
