+++ fastlane/metadata/android/de-DE/images/featureGraphic.jpg
