+++ fastlane/metadata/android/de-DE/images/icon.png
