+++ fastlane/metadata/android/fr-FR/images/featureGraphic.jpg
