+++ fastlane/metadata/android/fr-FR/images/icon.png
